+++ lib/shared/src/configuration.ts
@@ -19,6 +19,7 @@ export interface Configuration {
     autocompleteAdvancedAccessToken: string | null
     autocompleteAdvancedCache: boolean
     autocompleteAdvancedEmbeddings: boolean
+    autocompleteExperimentalTriggerMoreEagerly: boolean
     pluginsEnabled?: boolean
     pluginsDebugEnabled?: boolean
     pluginsConfig?: {